PIM-3882: Added a temporay file cleaner for mass edit
