+++ src/Pim/Bundle/BaseConnectorBundle/Resources/config/steps.yml
@@ -1,3 +1,3 @@
 parameters:
-    pim_base_connector.step.validator.class: Pim\Bundle\BaseConnectorBundle\Step\ValidatorStep
-    pim_base_connector.step.mass_edit_update_product_value.class: Pim\Bundle\BaseConnectorBundle\Step\MassEditProductStep
+    pim_base_connector.step.validator.class:                        Pim\Bundle\BaseConnectorBundle\Step\ValidatorStep
+    pim_base_connector.step.mass_edit.remove_temporary_media.class: Pim\Bundle\BaseConnectorBundle\Step\MassEditRemoveTemporaryMediaStep
