+++ src/Pim/Bundle/EnrichBundle/DependencyInjection/PimEnrichExtension.php
@@ -61,6 +61,7 @@ class PimEnrichExtension extends Extension
         $loader->load('readers.yml');
         $loader->load('processors.yml');
         $loader->load('savers.yml');
+        $loader->load('cleaners.yml');
 
         if ($config['record_mails']) {
             $loader->load('mail_recorder.yml');
