+++ src/Pim/Bundle/EnrichBundle/Resources/config/cleaners.yml
@@ -1,0 +1,11 @@
+parameters:
+    pim_enrich.mass_edit.remove_temporary_media.cleaner.class: Pim\Bundle\EnrichBundle\MassEditAction\Cleaner\EditCommonAttributesTemporaryFileCleaner
+
+services:
+    pim_enrich.mass_edit.remove_temporary_media.cleaner:
+        class: %pim_enrich.mass_edit.remove_temporary_media.cleaner.class%
+        arguments:
+            - '@pim_catalog.updater.product'
+            - '@pim_validator'
+            - '@pim_catalog.repository.product_mass_action'
+            - '@pim_catalog.repository.attribute'
